Fix libsanitizer linkage options for cross toolchains
In case of cross toolchain build, where GCC libraries are not installed
within the designated sysroot, the shared sanitizer libraries link
against libstdc++.so within the same directory. However this directory
is not in RPATH, so attempting to build a dynamically linked application
with -fsanitize=xx gives a linkage error. In the opinion of libsanitizer
maintainer, this is a problem with the tree structure of the
installation.

See: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=69839

Specifying an explicit -rpath-link is not a problem for the end-user,
but this problem also causes target checks for libasan regression tests
to fail. In order to be able to run regression tests on a cross
toolchain, this patch hacks the gcc driver and adds the requisite
-rpath-link option automatically for -fsanitize=address/undefined
relative to toolchain sysroot.

gcc/
        * gcc.cc (LIBSAN_RPATH): New macro.
        (LIBASAN_SPEC): Add LIBSAN_RPATH.
        (LIBUBSAN_SPEC): Likewise.
        (LIBTSAN_SPEC): Likewise.
        (LIBLSAN_SPEC): Likewise.

libsanitizer/
        * configure.ac (link_libsan_rpath): New config variable.
        * libsanitizer.spec.in (link_libsan_rpath): New spec.
        * configure: Regenerate with autoconf 2.69
        * Makefile.in: Regenerate with automake 1.15.1.
        * asan/Makefile.in: Likewise.
        * interception/Makefile.in: Likewise.
        * libbacktrace/Makefile.in: Likewise.
        * lsan/Makefile.in: Likewise.
        * sanitizer_common/Makefile.in: Likewise.
        * tsan/Makefile.in: Likewise.
        * ubsan/Makefile.in: Likewise.
        * hwasan/Makefile.in: Likewise.

Cherry-picked e45e8831afb40eddfa82bedf7fc5536619d7382f,
25c48b961e16748cfa468830040a865ece7dd235 and
96b7557473a09bb181be86c741c6e79a4a3c41b1
from https://github.com/MIPS/gcc

Signed-off-by: Faraz Shahbazker <fshahbazker@wavecomp.com>
Signed-off-by: Chao-ying Fu <cfu@mips.com>
Signed-off-by: Aleksandar Rakic <aleksandar.rakic@htecgroup.com>